# feat: improve performance by pre-compiling regexes