This PR minimizes the usage of `regexp.MustCompile` by precompiling the static ones and using a cache for dynamic ones.