@surik 2024-09-15
@hsluoyz  I did a simple benchmark test, please see [gist](https://gist.github.com/surik/27308bf52ff8f2fddd6a54210b0f5594)/

Results:
```
goos: darwin
goarch: arm64
pkg: github.com/casbin/casbin/v2/util
BenchmarkNewKeyMatch2-10          359047              3282 ns/op
BenchmarkOldKeyMatch2-10          287622              4077 ns/op
BenchmarkOldKeyGet2-10            222024              5476 ns/op
BenchmarkNewKeyGet2-10           1000000              1043 ns/op
```