@surik 2024-09-15
@hsluoyz `Old` is before and `New` with the changes. It also can be tested like this:
``` go
package util

import (
	"testing"
)

func BenchmarkKeyMatch2(b *testing.B) {
	for i := 0; i < b.N; i++ {
		KeyMatch2("/myid/using/myresid", "/:id/using/:resId")
	}
}

func BenchmarkKeyGet2(b *testing.B) {
	for i := 0; i < b.N; i++ {
		KeyGet2("/myid/using/myresid", "/:id/using/:resId", "resId")
	}
}
```

Before the changes:
```
➜  casbin git:(master) ✗ go test -bench=. ./util
goos: darwin
goarch: arm64
pkg: github.com/casbin/casbin/v2/util
BenchmarkKeyMatch2-10             298600              4029 ns/op
BenchmarkKeyGet2-10               226227              5308 ns/op
```

After the changes:
```
➜  casbin git:(speedup-regexp) ✗ go test -bench=. ./util
goos: darwin
goarch: arm64
pkg: github.com/casbin/casbin/v2/util
BenchmarkKeyMatch2-10             364015              3272 ns/op
BenchmarkKeyGet2-10              1000000              1083 ns/op
```

In this simple benchmark `KeyMatch2` is 20% faster and `KeyGet2` is 80% faster than the implementations before the change.

Please, let me know if you want to see a more specific comparison.