@ALX99 2024-10-01
Does this not cause a memory leak? The cache can only grow but never shrink